Split files in `pull_request` remediation golang package
This _hopefully_ makes these easier to browse and denotes where things should be.

`internal.go` didn't tell much to the reader.
